src/requirements.txt
fastapi==0.112.1
uvicorn==0.30.6
python-dotenv==1.0.1
SQLAlchemy==2.0.32
gunicorn==20.1.0
psycopg2-binary==2.9.9
loguru==0.7.0
requests==2.32.3
PyJWT==2.9.0
azure.identity==1.14.0  
azure-keyvault-secrets==4.7.0
#pycrypto==2.6.1
pycryptodome==3.20.0
rsa==4.9
python-jose==3.3.0
pyodbc==5.1.0
httpx==0.27.0